fix: 401 for bots fix a critical error where bots would 401 due to incorrect headers.
while revolt's documentation shows X-Session-Token as the auth header, X-Bot-Token is required instead. Both will be set as the library does not know if it's acting as a self-bot or not.

idea: in the future, maybe during the initial handshake we could fetch the user, then the request method can determine if we're a self-bot by checking s.State.User?