Use distinct event update types over AbstractEventUpdate
This is a massive overhaul to the event system; mergeJSON is no longer used for updating existing memory objects. Instead, there are direct methods such as .update() and .clear() for those structs.

Additionally, handlers are now added to a map instead of distinct slices.

AddHandler's signature has been completely revamped and it now uses generics to infer and validate the event type, then add it to the handlers map.